feat(theme): move locales into a dedicated folder
